Build for netcoreapp3.0. Get the package building.
