+++ src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf8Utility.WhiteSpace.CoreLib.cs
@@ -2,6 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Runtime.InteropServices;
+using Internal.Runtime.CompilerServices;
+
+#pragma warning disable SA1121 // explicitly using type aliases instead of built-in types
+#if TARGET_64BIT
+using nint = System.Int64;
+using nuint = System.UInt64;
+#else
+using nint = System.Int32;
+using nuint = System.UInt32;
+#endif
+
 namespace System.Text.Unicode
 {
     internal static partial class Utf8Utility
@@ -12,24 +24,28 @@ namespace System.Text.Unicode
         /// </summary>
         public static int GetIndexOfFirstNonWhiteSpaceChar(ReadOnlySpan<byte> utf8Data)
         {
+            return (int)GetIndexOfFirstNonWhiteSpaceChar(ref MemoryMarshal.GetReference(utf8Data), (uint)utf8Data.Length);
+        }
+
+        private static nuint GetIndexOfFirstNonWhiteSpaceChar(ref byte utf8Data, nuint length)
+        {
             // This method is optimized for the case where the input data is ASCII, and if the
             // data does need to be trimmed it's likely that only a relatively small number of
             // bytes will be trimmed.
 
-            int i = 0;
-            int length = utf8Data.Length;
+            nuint i = 0;
 
             while (i < length)
             {
                 // Very quick check: see if the byte is in the range [ 21 .. 7F ].
                 // If so, we can skip the more expensive logic later in this method.
 
-                if (utf8Data[i] > (sbyte)0x20)
+                if ((sbyte)Unsafe.AddByteOffset(ref utf8Data, i) > (sbyte)0x20)
                 {
                     break;
                 }
 
-                uint possibleAsciiByte = utf8Data[i];
+                uint possibleAsciiByte = Unsafe.AddByteOffset(ref utf8Data, i);
                 if (UnicodeUtility.IsAsciiCodePoint(possibleAsciiByte))
                 {
                     // The simple comparison failed. Let's read the actual byte value,
@@ -47,10 +63,10 @@ namespace System.Text.Unicode
                     // Not ASCII data. Go back to the slower "decode the entire scalar"
                     // code path, then compare it against our Unicode tables.
 
-                    Rune.DecodeFromUtf8(utf8Data.Slice(i), out Rune decodedRune, out int bytesConsumed);
+                    Rune.DecodeFromUtf8(new ReadOnlySpan<byte>(ref utf8Data, (int)length).Slice((int)i), out Rune decodedRune, out int bytesConsumed);
                     if (Rune.IsWhiteSpace(decodedRune))
                     {
-                        i += bytesConsumed;
+                        i += (uint)bytesConsumed;
                         continue;
                     }
                 }
@@ -68,23 +84,26 @@ namespace System.Text.Unicode
         /// </summary>
         public static int GetIndexOfTrailingWhiteSpaceSequence(ReadOnlySpan<byte> utf8Data)
         {
+            return (int)GetIndexOfTrailingWhiteSpaceSequence(ref MemoryMarshal.GetReference(utf8Data), (uint)utf8Data.Length);
+        }
+
+        private static nuint GetIndexOfTrailingWhiteSpaceSequence(ref byte utf8Data, nuint length)
+        {
             // This method is optimized for the case where the input data is ASCII, and if the
             // data does need to be trimmed it's likely that only a relatively small number of
             // bytes will be trimmed.
-
-            int length = utf8Data.Length;
 
             while (length > 0)
             {
                 // Very quick check: see if the byte is in the range [ 21 .. 7F ].
                 // If so, we can skip the more expensive logic later in this method.
 
-                if ((sbyte)utf8Data[length - 1] > (sbyte)0x20)
+                if ((sbyte)Unsafe.Add(ref Unsafe.AddByteOffset(ref utf8Data, length), -1) > (sbyte)0x20)
                 {
                     break;
                 }
 
-                uint possibleAsciiByte = utf8Data[length - 1];
+                uint possibleAsciiByte = Unsafe.Add(ref Unsafe.AddByteOffset(ref utf8Data, length), -1);
                 if (UnicodeUtility.IsAsciiCodePoint(possibleAsciiByte))
                 {
                     // The simple comparison failed. Let's read the actual byte value,
@@ -102,10 +121,10 @@ namespace System.Text.Unicode
                     // Not ASCII data. Go back to the slower "decode the entire scalar"
                     // code path, then compare it against our Unicode tables.
 
-                    Rune.DecodeLastFromUtf8(utf8Data.Slice(length), out Rune decodedRune, out int bytesConsumed);
+                    Rune.DecodeLastFromUtf8(new ReadOnlySpan<byte>(ref utf8Data, (int)length), out Rune decodedRune, out int bytesConsumed);
                     if (Rune.IsWhiteSpace(decodedRune))
                     {
-                        length -= bytesConsumed;
+                        length -= (uint)bytesConsumed;
                         continue;
                     }
                 }
