+++ src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf8Utility.WhiteSpace.NonCoreLib.cs
@@ -2,18 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Runtime.InteropServices;
-using Internal.Runtime.CompilerServices;
-
-#pragma warning disable SA1121 // explicitly using type aliases instead of built-in types
-#if TARGET_64BIT
-using nint = System.Int64;
-using nuint = System.UInt64;
-#else
-using nint = System.Int32;
-using nuint = System.UInt32;
-#endif
-
 namespace System.Text.Unicode
 {
     internal static partial class Utf8Utility
@@ -24,35 +12,31 @@ namespace System.Text.Unicode
         /// </summary>
         public static int GetIndexOfFirstNonWhiteSpaceChar(ReadOnlySpan<byte> utf8Data)
         {
-            return (int)GetIndexOfFirstNonWhiteSpaceChar(ref MemoryMarshal.GetReference(utf8Data), (uint)utf8Data.Length);
-        }
-
-        private static nuint GetIndexOfFirstNonWhiteSpaceChar(ref byte utf8Data, nuint length)
-        {
             // This method is optimized for the case where the input data is ASCII, and if the
             // data does need to be trimmed it's likely that only a relatively small number of
             // bytes will be trimmed.
 
-            nuint i = 0;
+            int i = 0;
+            int length = utf8Data.Length;
 
             while (i < length)
             {
                 // Very quick check: see if the byte is in the range [ 21 .. 7F ].
                 // If so, we can skip the more expensive logic later in this method.
 
-                if ((sbyte)Unsafe.AddByteOffset(ref utf8Data, i) > (sbyte)0x20) //TODO: remove IntPtr cast
+                if (utf8Data[i] > (sbyte)0x20)
                 {
                     break;
                 }
 
-                uint possibleAsciiByte = Unsafe.AddByteOffset(ref utf8Data, i); //TODO: remove IntPtr cast
+                uint possibleAsciiByte = utf8Data[i];
                 if (UnicodeUtility.IsAsciiCodePoint(possibleAsciiByte))
                 {
                     // The simple comparison failed. Let's read the actual byte value,
                     // and if it's ASCII we can delegate to Rune's inlined method
                     // implementation.
 
-                    if (Rune.IsWhiteSpace(Rune.UnsafeCreate(possibleAsciiByte)))
+                    if (Rune.IsWhiteSpace(new Rune(possibleAsciiByte)))
                     {
                         i++;
                         continue;
@@ -63,10 +47,10 @@ namespace System.Text.Unicode
                     // Not ASCII data. Go back to the slower "decode the entire scalar"
                     // code path, then compare it against our Unicode tables.
 
-                    Rune.DecodeFromUtf8(new ReadOnlySpan<byte>(ref utf8Data, (int)length).Slice((int)i), out Rune decodedRune, out int bytesConsumed);
+                    Rune.DecodeFromUtf8(utf8Data.Slice(i), out Rune decodedRune, out int bytesConsumed);
                     if (Rune.IsWhiteSpace(decodedRune))
                     {
-                        i += (uint)bytesConsumed;
+                        i += bytesConsumed;
                         continue;
                     }
                 }
@@ -84,33 +68,30 @@ namespace System.Text.Unicode
         /// </summary>
         public static int GetIndexOfTrailingWhiteSpaceSequence(ReadOnlySpan<byte> utf8Data)
         {
-            return (int)GetIndexOfTrailingWhiteSpaceSequence(ref MemoryMarshal.GetReference(utf8Data), (uint)utf8Data.Length);
-        }
-
-        private static nuint GetIndexOfTrailingWhiteSpaceSequence(ref byte utf8Data, nuint length)
-        {
             // This method is optimized for the case where the input data is ASCII, and if the
             // data does need to be trimmed it's likely that only a relatively small number of
             // bytes will be trimmed.
+
+            int length = utf8Data.Length;
 
             while (length > 0)
             {
                 // Very quick check: see if the byte is in the range [ 21 .. 7F ].
                 // If so, we can skip the more expensive logic later in this method.
 
-                if ((sbyte)Unsafe.Add(ref Unsafe.AddByteOffset(ref utf8Data, (IntPtr)length), -1) > (sbyte)0x20) //TODO: remove IntPtr cast
+                if ((sbyte)utf8Data[length - 1] > (sbyte)0x20)
                 {
                     break;
                 }
 
-                uint possibleAsciiByte = Unsafe.Add(ref Unsafe.AddByteOffset(ref utf8Data, (IntPtr)length), -1); //TODO: remove IntPtr cast
+                uint possibleAsciiByte = utf8Data[length - 1];
                 if (UnicodeUtility.IsAsciiCodePoint(possibleAsciiByte))
                 {
                     // The simple comparison failed. Let's read the actual byte value,
                     // and if it's ASCII we can delegate to Rune's inlined method
                     // implementation.
 
-                    if (Rune.IsWhiteSpace(Rune.UnsafeCreate(possibleAsciiByte)))
+                    if (Rune.IsWhiteSpace(new Rune(possibleAsciiByte)))
                     {
                         length--;
                         continue;
@@ -121,10 +102,10 @@ namespace System.Text.Unicode
                     // Not ASCII data. Go back to the slower "decode the entire scalar"
                     // code path, then compare it against our Unicode tables.
 
-                    Rune.DecodeLastFromUtf8(new ReadOnlySpan<byte>(ref utf8Data, (int)length), out Rune decodedRune, out int bytesConsumed);
+                    Rune.DecodeLastFromUtf8(utf8Data.Slice(length), out Rune decodedRune, out int bytesConsumed);
                     if (Rune.IsWhiteSpace(decodedRune))
                     {
-                        length -= (uint)bytesConsumed;
+                        length -= bytesConsumed;
                         continue;
                     }
                 }
