+++ src/libraries/System.Private.CoreLib/src/System/Text/Utf8Span.Manipulation.cs
@@ -431,7 +431,11 @@ namespace System.Text
 
                         if (SearchRune >= 0)
                         {
+#if SYSTEM_PRIVATE_CORELIB || NETSTANDARD2_0
                             wasMatchFound = searchSpan.TryFind(Rune.UnsafeCreate((uint)SearchRune), out matchRange);
+#else
+                            wasMatchFound = searchSpan.TryFind(new Rune((uint)SearchRune), out matchRange);
+#endif
                         }
                         else
                         {
