+++ src/libraries/System.Utf8String.Experimental/pkg/System.Utf8String.Experimental.pkgproj
@@ -3,7 +3,7 @@
   <Import Project="$([MSBuild]::GetPathOfFileAbove(Directory.Build.props))" />
   <ItemGroup>
     <ProjectReference Include="..\ref\System.Utf8String.Experimental.csproj">
-        <SupportedFramework>netcoreapp5.0;</SupportedFramework>
+      <SupportedFramework>net461;netcoreapp2.0;uap10.0.16299;$(AllXamarinFrameworks)</SupportedFramework>
     </ProjectReference>
     <ProjectReference Include="..\src\System.Utf8String.Experimental.csproj" />
   </ItemGroup>
