+++ src/libraries/System.Utf8String.Experimental/src/Resources/Strings.resx
@@ -147,6 +147,9 @@
   <data name="Utf8Span_CannotCallEqualsObject" xml:space="preserve">
     <value>Cannot call Utf8Span.Equals(object). Use Equals(Utf8Span) or operator == instead.</value>
   </data>
+  <data name="Utf8String_CallbackProvidedMalformedData" xml:space="preserve">
+    <value>The callback populated its buffer with ill-formed UTF-8 data. Callbacks are required to populate the buffer only with well-formed UTF-8 data.</value>
+  </data>
   <data name="Utf8String_CannotSplitMultibyteSubsequence" xml:space="preserve">
     <value>Cannot create the desired substring because it would split a multi-byte UTF-8 subsequence.</value>
   </data>
